Update backend endpoints, migrations and DbContext

+++ backend/eParking/Endpoints/ParkingSpotTypeEndpoints/ParkingSpotTypeGetAllEndpoint.cs
@@ -50,7 +50,7 @@ public class ParkingSpotTypesGetAllEndpoint(ApplicationDbContext db) : MyEndpoin
         public int ID { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; }
-        public int PriceMultiplier { get; set; }
+        public decimal PriceMultiplier { get; set; }
     }
 }
